requirements.txt
requests
pycryptodome
cookiecutter
